GUI adjustments
The timing in order to locate the source picture in its album of a linked
picture in another album has been reasonably adjusted. The home button
(preliminary: ⌂ symbol) was lowered a little to fit within the window when
the new location of the #highUP position (now includes the very first
album header). Whitespace nowrap in the texts below miniature has been
dropt, since it produces added white aread to the right when they are dragged.