Spinner
The spinner behaviour was adjusted. Async/await seems to be problematic when
the interacting functions are located within different scopes, e.g.
this.actions.fu1(), this.fu2(), and just fu3() -- maybe needs special
consideration in order to combine well. Thus such use for spinner timing was
abandoned throughapproximations with pausing, etc. Not ideal but handable.