System upgrade
Upgraded Kubuntu to 20.04 and also Node etc. to latest or near
latest. All is now NVM installs except that a symlink to the actual
Node is manually maintained in /usr/bin for use of Node in e.g.
start scripts and other cases. A lot of depreciation warnings are
still to be dealt with. Starting transition to better-sqlite3.